# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface OwnerStateUpdate {
  /** @minimum 0 */
  expectedRevision: number;
  snapshot: unknown;
}

export interface OwnerStateResponse {
  initialized: boolean;
  revision?: number;
  snapshot?: unknown;
  updatedAt?: string;
}

export interface UploadUrlRequest {
  /**
     * @minLength 1
     * @maxLength 255
     */
  name: string;
  /**
     * @minimum 1
     * @maximum 12582912
     */
  size: number;
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata: UploadUrlRequest;
}

export interface UploadCompleteRequest {
  objectPath: string;
}

export interface UploadCompleteResponse {
  objectPath: string;
  url: string;
}