# lib/api-zod/src/generated/types/ownerStateResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OwnerStateResponse {
  initialized: boolean;
  revision?: number;
  snapshot?: unknown;
  updatedAt?: Date;
}